Fix int64 type compatibility for DuckDB
- Changed Cidr struct fields from int to int64
- Updated IPv4toDecimal function to return int64
- Fixed cidrObject struct to use int64 for IP decimals
- This ensures full IPv4 address range support (0-4,294,967,295)

+++ pkg/model/cidr.go
@@ -7,8 +7,8 @@ import (
 
 type Cidr struct {
     Net string
-	Start_ip      int
-	End_ip        int
+	Start_ip      int64
+	End_ip        int64
 	Url           string
 	Cloudplatform string
 	Iptype        string
